Refactor: move LineListsWindow class to its own .py file.
